+++ lib/src/domain/user_options/models/output_mode.dart
@@ -1,0 +1,4 @@
+enum OutputMode {
+  cli,
+  markdown,
+}