+++ lib/src/extensions/string.dart
@@ -1,0 +1,8 @@
+extension RemovePrefix on String {
+  String removePrefix(String prefix) {
+    if (startsWith(prefix)) {
+      return replaceFirst(prefix, "");
+    }
+    return this;
+  }
+}