+++ lib/src/presentation/output_print_generator/output_generator.dart
@@ -1,3 +1,5 @@
+import 'package:pull_request_coverage/src/domain/analyser/models/analysis_result.dart';
+
 abstract class OutputGenerator {
   String? getFileHeader(String filePath, int uncoveredLinesCount, int totalNewLinesCount);
 
@@ -6,4 +8,6 @@ abstract class OutputGenerator {
   String? getLine(String line, int lineNumber, bool isANewLine, bool isCovered);
 
   String? getSourceCodeFooter();
+
+  String? getResume(AnalysisResult analysisResult, double? minimumCoverageRate, int? maximumUncoveredLines);
 }